Attempt to fix 3d pop up shape
Tried various methods of fixing the 3d pop up shape that Bernie started. No luck.
I have managed to get the shape to face the camera no matter what the camera position, but this is not helpful by itself
Still need to move the object to be centered where the camera is looking (ie, move the object). Could possibly do this using the camera position and directional normal and move
the centre of the shape to that XY position, but this has the problem of needing to update the other 8 vertices in relation to the new centre vertex.